FIX: Negative scale not working in IKFK2Bones.

+++ src/nodesLib/ldRigFkIk2Bones.h
@@ -19,7 +19,7 @@ public:
     void computeIk(MVector rootPos, MVector upPos, MVector effPos,
                     double b1Length, double b2Length,
                     int activeStretch, int alignAxis, int upVectorAxis,
-                    double negativeScale, double invertIk,
+                    int negativeScale, double invertIk,
                     MTransformationMatrix &b1Trans, MTransformationMatrix &b2Trans, MTransformationMatrix &b3Trans);
     MTransformationMatrix blendTransform(MTransformationMatrix matrixA, MTransformationMatrix matrixB, double blend);
     MTransformationMatrix buildTransform(MVector pos, MVector axis0, MVector axis1,
